Update README.md
Rename/repurpose repo to be more generic.

Expand on goal of the repo and the future contents. This is going to be
used as a sandbox for various tools to automate away tasks I've found
tedious in the past.

I try and upstream improvements where possible, but this repo generally
contains items which don't make sense to upstream.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>